+++ frontend/src/handlers/user/imageUpdateHandler.js
@@ -1,6 +1,6 @@
-import fetchAPI from "../utils/fetchAPI.js";
-import { myInfoStore } from "../store/initialStates.js";
-import showToast from "../utils/showToast.js";
+import fetchAPI from "../../utils/fetchAPI.js";
+import { myInfoStore } from "../../store/initialStates.js";
+import showToast from "../../utils/showToast.js";
 
 const imageUpdateHandler = (id) => {
   const fileInput = document.createElement("input");
